backend/requirements.txt
Flask==2.1.2
pymysql==1.0.2
Werkzeug==2.1.2
requests==2.25.1
psycopg2-binary>=2.8
flask_cors == 3.0.10
bcrypt >= 2.0.0
python-dotenv == 1.0.0
flask_mail == 0.7.6
PyJWT == 2.1.0